pyproject.toml
[project]
name = "imdb-reviews-sentiment-classifier"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dvc>=3.64.2",
    "fire>=0.7.1",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "mlflow==3.8.0",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[dependency-groups]
dev = [
    "dvc-s3>=3.2.2",
    "pre-commit>=4.5.0",
    "ruff>=0.14.9",
]

[tool.ruff]
line-length = 100
extend-exclude = [".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
docstring-code-format = true
quote-style = "double"
indent-style = "space"
